+++ chatbot/backend/_OLDGROUP/main_OLDGROUP.py
@@ -1,5 +1,5 @@
 from database_OLDGROUP.chroma_db import get_global_collection
-from chatbot.backend.server_OLDGROUP import create_server
+from chatbot.backend.OLDGROUP.server_OLDGROUP import create_server
 
 # Creates backend server and initialises vector database
 app = create_server()
