+++ chatbot/backend/services/models/embedding_model.py
